Wildlife-Guardian/streamlit_model/requirements.txt
streamlit==1.40.1
Pillow==10.0.0
tensorflow==2.15.0  # Uncomment when you add the exported Teachable Machine model
